[OVN] Change ControllerAgent type dinamically
The ControllerAgent class now changes in Openstack when changing the
type of an agent in its ovsdb table. This is possible thanks to adding a
new condition to the event that triggered the update of the AgentCache.
If the right external_ids changed, a new agent with the correct class
will replace the old one in the AgentCache. Also adds functional
tests to check this.

Closes-Bug: #1934420

Signed-off-by: Elvira García <egarciar@redhat.com>
Change-Id: I75380c455864aa055e8049a6beff8619687b52ce
(cherry picked from commit 9cffd7b3223c2e8c88129c79aa62c255350c55d4)