Revert "Config option to enable OVN IDL on other workers"
This reverts commit 66ac943b6435901c163c3050ad41dfcaf36ac11d.

Reason for revert: After discussions with the core reviewers for ML2/OVN
and Neutron folks about this configuration option we understood that we
should just enable RpcWorkers  to connect to the OVN database by default
and later allow rpc_workers to be set to 0 in Neutron (for those looking
for optimizations and that don't need RpcWorkers in their deployment). So
I am proposing a revert for this configuration option.

Change-Id: I73ba195401fff9f832b84633398816383f710b23