[OVN] Fix ML2/OVN + Neutron DHCP agent use case
In an attempt to minimize the number of connections to the OVSDBs the
ML2/OVN driver stopped connecting to the OVN dbs if the Neutron worker
was a RpcWorker but, that introduced a regression for those using
ML2/OVN + Neutron DHCP agent (for baremetal provisioning).

Because the RpcWorker didn't have a connection to the OVN database the
Neutron DHCP agent wasn't able to create the DHCP port and failed with:

2021-07-13 13:36:18.116 17 ERROR neutron.plugins.ml2.managers if not
self._sb_ovn.chassis_exists(host):
2021-07-13 13:36:18.116 17 ERROR neutron.plugins.ml2.managers
AttributeError: 'NoneType' object has no attribute 'chassis_exists'

This patch adds the RpcWorker back to the list of workers that should
connect to the OVN databases.

Change-Id: I454f3b52376a02f6cc9ca9124083295631bd03d5
Closes-Bug: #1929633
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>