Allow ovn_db_sync to continue on duplicate normalised CIDR
OVN now uses normalised CIDR when adding a security group rule[1]. It
uses may_exist=True for adding ACL (secgroup rule), in case there are
multiple CIDRs in neutron that normalises to the same.

Do the same in ovn_db_sync, so that the sync don't fail hard on such
duplicates.

[1] https://review.opendev.org/c/openstack/neutron/+/736386/

Change-Id: I9d9c21e460029e4a6a845520bfcc2889ad20429b
Related-Bug: #1869129
Closes-Bug: #1961112