Promote neutron-ovn-tempest-slow job to be voting and gating
This jobs is almost the same as tempest-slow-py3 since we switched
OVN to be default backend in Neutron. And that tempest-slow-py3 job
is used by many projects. So to avoid potential breaks of the gate for
other projects (like we did recently, see related bug for details)
let's make this job voting and gating.
As it is really used in many different projects as voting and gating
job already I don't think there is any issue with doing the same in
the Neutron gate.

Related-bug: #1936983
Related-bug: #1930402

Change-Id: I85d3830e9cc65162db846e4858871e1db547a04b