ovn: Don't fail db-sync if port binding changes
During migration from OVS to OVN it can happen that gateway ports are
scheduled to a different gateway chassis when Neutron is running. This
patch doesn't fail in such case. The migration procedure runs the db
sync twice in a row so it should be good to not perform any action when
this happens and let the next migration handle that.

Change-Id: I28a4a5fef20d5049f4887d43006947b434de3d78
Closes-Bug: #1939704
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>