Fix gate for neutron-lib v2.14
A patch [1] that updates upper-constraints.txt for neutron-lib 2.14 fails
to pass check pipeline.

Due to development sequence it's necessary to append a new rule type
to VALID_RULE_TYPES in Neutron first, and then move it to neutron-lib.
Because of this process, there's a risk that the new rule type is
present in both Neutron and in neutron-lib.

This is why check pipeline keeps failing with neutron-lib v2.14, as it
contains 'packet_rate_limit' and in Neutron we append it to the list
anyways, ending up with duplicates.

This patch ensures that there are no duplicates in VALID_RULE_TYPES.

[1] https://review.opendev.org/c/openstack/requirements/+/805352

Change-Id: Ib6963f402c9fec8169afcf467d613bba4e06130d