Fix typos in agent policy names
I think these typos were an accident, but it's causing the generated
policy files (oslopolicy-sample-generator --namespace neutron) to alias
the old rule name with the new rule name because it thinks we're change
them to something else.

This commit corrects them to the old values since I don't think the
names are changing.

Change-Id: I615c03567a4e96293f4138f4f171ae0fb09dc2f1