update subport status when trunk/subport create/delete is triggerred
When trunk or subport is created or deleted, notifications are sent. In
the meanwhile, subport status is not updated accordingly.

device_id device_owner and binding:host_id should be updated when
notifications are received.

In this patch, notifications are listened and subport status should be
updated when trunk/subport is created and subport is already added to
trunk:

    device_id: <trunk_id>
    device_owner: trunk:subport
    binding:host_id: <parent port host_id>

When trunk/subport is deleted or subport is removed from trunk, subport
status should be updated:

    device_id: None
    device_owner: None
    binding:host_id: None

Closes-Bug: #1942413
Change-Id: I0bdc5aabae0125926253b530fd0c6e79ba7bcbb9