[DVR] Check if SNAT iptables manager is initialized
Check if SNAT iptables manager is initialized before processing the
IP NAT rules. If the router never had an external GW port, the DVR
GW in the SNAT namespace has not been created and the SNAT iptables
manager has not been initialized.

In this case, the IP NAT rules for centralized FIPs (to be applied
on the SNAT namespace) cannot be set.

Closes-Bug: #1945215
Change-Id: I426602514805d728f8cd78e42f2b0979b2101089