Execute "migrate_neutron_database_to_ovn" inside the same DB ctx
All DB commands executed in "migrate_neutron_database_to_ovn"
should be executed inside the same write context. That will prevent
the DB transaction errors reported in the BZ. When a OVO object is
retrieved and updated, both DB commands should be executed in the
same DB transaction to ensure the DB object stored in the OVO
(that is a stateful representation of the requested columns) is
fresh and not modified by other transaction.

Related-Bug: #1939704
Change-Id: I754e2f9ad5b896fcc14ff30bd2901ccec2471907