Add functional and fullstack jobs with FIPS enabled
This patch adds definition of the functional and fullstack jobs
with enabled support for FIPS [1].
Jobs are based on the Centos 8 stream as this disto allows to enable
FIPS support.
Jobs are added to the experimental queue for now.

This patch also makes some changes in the bindep and
configure_functional_tests role to make functional/fullstack tests
working on the Centos.

[1] https://csrc.nist.gov/publications/detail/fips/140/3/final

Co-Authored-By: Ade Lee <alee@redhat.com>
Change-Id: I582495826155740ad2660ee2a8717696b0393d26