[OVN] Execute OVN migration transactions independently
Instead of executing all Neutron DB changes in one single
transaction, in order to avoid clashes with other DB operations
each element will be treated independently.

For the "PortBinding" registers update, each element will be updated
in an independent transaction.

For the "Trunk" registers update, all subports "PortBinding"
registers belonging to this trunk will be updated in an independent
transaction.

Closes-Bug: #1939704
Change-Id: I6f78750af34001b3707f1a6ee3a0882b427a5dcc