[OVS] Workaround when OpenFlow controller restarts
When a new OF controller is added or when new OF protocols are added to
a bridge, the OF controller is restarted. When that happens, the os-ken
``OfctlService`` instance removes the OF controller datapath and waits
for an update event that will add a new one. During this time no OF
controllers (datapaths) will be registered and any request will be
rejected with a ``InvalidDatapath`` exception.

This patch is a temporary workaround that prevents any command from
being executed just after those two described operations, providing
time to OVS to send the OF updates.

Conflicts:
	neutron/agent/common/ovs_lib.py

Related-Bug: #1948642
Change-Id: I03878a2e650d5f0d0167e0f7daa351b12bae8725
(cherry picked from commit 0194856da19f0ce96b613968ab415821037f18d0)