Don't enforce scopes in the API policies UT temporary
There is bug in oslo_policy [1] that caused issues with enforcing
scopes properly for rules which inherits from the oslo_policy.BaseCheck
class (many of Neutron rules are done like that).
Now when fix [2] will be merged in oslo_policy we will need to change
expected exception type in some of the policies UT from
PolicyNotAuthorized to InvalidScope.
For now, to make it working with both old and new oslo_policy, let's
disable enforcing scopes in those UT. It will be enabled again when new
oslo_policy will be fixed and we will us it in Neutron.

[1] https://launchpad.net/bugs/1923503
[2] https://review.opendev.org/c/openstack/oslo.policy/+/804980

Related-Bug: #1923503
Change-Id: I5c64eb9315b37379b0899c9d35f2f0d96f9c1579