[OVN] Update check_for_mcast_flood_reports() to check for mcast_flood
The patch [0] changed the value of the "mcast_flood" option for the
localnet ports in OVN to "false" to avoid multicast packet duplication
but, the maintenance task check_for_mcast_flood_reports() was never
updated to account for that change.

So, localnet ports created prior to [0] would continue to have the
"mcast_flood" value as "true" and new ports would have it as "false".
This patch fixes the maintenance task.

[0] https://review.opendev.org/c/openstack/neutron/+/797418

Change-Id: Ib90704a85c9ed659d71ac8ed42b1b542bd2496a2
Closes-Bug: #1949081
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>