[OVN] Handle OVN agents when "Chassis" register is deleted
If an "ovn-controller" ends not gracefully, the node "Chassis" and
"Chassis_Private" registers will remain in the OVN SB database.
Because there is no a mandatory procedure to delete the "Chassis"
and "Chassis_Private" registers, the administrator can manually
delete, from the OVN SB database, any register in any order.

If the "Chassis" register is deleted and the Neutron server restarted,
the updated "Chassis_Private" register will be read from the database.
That won't contain the "Chassis" information as this register has been
deleted. In this case, the ``NeutronAgent`` returns ``DeletedChassis``,
an empty chassis register with no information.

NOTE: the sequence of actions ("Chassis" register deletion, Neutron
server restart) must be follow to reproduce this issue. If the
"Chassis" register is deleted, the Neutron server OVN agent local cache
won't update the stored information and will keep the previous value.
It is when the Neutron server is restarted when the OVN agent local
cache is retrieved again; at this time the "Chassis_Private" register
won't have any related "Chassis" register.

Closes-Bug: #1951149

Change-Id: I17aa53cea6aba8ea83187c99102a6f25fd33cfff