Ignore warnings in the policies UT
Currently we are not enforcing scopes in the policy checks in unit
tests to avoid issues after [1] will be merged.
But still scope configured in the test contexts and in the rules don't
match so oslo.policy produces a lot of warnings and it seems in the
results of [1] that this may be too much for stestr and UT job is
timing out there.
So to avoid that, let's ignore python warnings in the policy related
unit tests.

[1] https://review.opendev.org/c/openstack/oslo.policy/+/804980

Related-Bug: #1923503
Change-Id: I3264f0708c7899bf886c6b2c14d8f80dae337e8d